+++ tenantforge/migrations/0013_evidence_bundles.down.sql
@@ -1,0 +1,1 @@
+DROP TABLE IF EXISTS tf_evidence_bundles;